Move departements.py in utils/ Fix #119
